+++ app/scripts/popup.js
@@ -10,69 +10,75 @@ const NotificationManager = require('./lib/notification-manager')
 const notificationManager = new NotificationManager()
 const setupRaven = require('./lib/setupRaven')
 
-// create platform global
-global.platform = new ExtensionPlatform()
+start().catch(log.error)
 
-// setup sentry error reporting
-const release = global.platform.getVersion()
-setupRaven({ release })
+async function start() {
 
-// inject css
-// const css = MetaMaskUiCss()
-// injectCss(css)
+  // create platform global
+  global.platform = new ExtensionPlatform()
 
-// identify window type (popup, notification)
-const windowType = isPopupOrNotification()
-global.METAMASK_UI_TYPE = windowType
-closePopupIfOpen(windowType)
+  // setup sentry error reporting
+  const release = global.platform.getVersion()
+  setupRaven({ release })
 
-// setup stream to background
-const extensionPort = extension.runtime.connect({ name: windowType })
-const connectionStream = new PortStream(extensionPort)
+  // inject css
+  // const css = MetaMaskUiCss()
+  // injectCss(css)
 
-// start ui
-const container = document.getElementById('app-content')
-startPopup({ container, connectionStream }, (err, store) => {
-  if (err) return displayCriticalError(err)
+  // identify window type (popup, notification)
+  const windowType = isPopupOrNotification()
+  global.METAMASK_UI_TYPE = windowType
+  closePopupIfOpen(windowType)
 
-  // Code commented out until we begin auto adding users to NewUI
-  // const { isMascara, identities = {}, featureFlags = {} } = store.getState().metamask
-  // const firstTime = Object.keys(identities).length === 0
-  const { isMascara, featureFlags = {} } = store.getState().metamask
-  let betaUIState = featureFlags.betaUI
+  // setup stream to background
+  const extensionPort = extension.runtime.connect({ name: windowType })
+  const connectionStream = new PortStream(extensionPort)
 
-  // Code commented out until we begin auto adding users to NewUI
-  // const useBetaCss = isMascara || firstTime || betaUIState
-  const useBetaCss = isMascara || betaUIState
+  // start ui
+  const container = document.getElementById('app-content')
+  startPopup({ container, connectionStream }, (err, store) => {
+    if (err) return displayCriticalError(err)
 
-  let css = useBetaCss ? NewMetaMaskUiCss() : OldMetaMaskUiCss()
-  let deleteInjectedCss = injectCss(css)
-  let newBetaUIState
+    // Code commented out until we begin auto adding users to NewUI
+    // const { isMascara, identities = {}, featureFlags = {} } = store.getState().metamask
+    // const firstTime = Object.keys(identities).length === 0
+    const { isMascara, featureFlags = {} } = store.getState().metamask
+    let betaUIState = featureFlags.betaUI
 
-  store.subscribe(() => {
-    const state = store.getState()
-    newBetaUIState = state.metamask.featureFlags.betaUI
-    if (newBetaUIState !== betaUIState) {
-      deleteInjectedCss()
-      betaUIState = newBetaUIState
-      css = betaUIState ? NewMetaMaskUiCss() : OldMetaMaskUiCss()
-      deleteInjectedCss = injectCss(css)
-    }
-    if (state.appState.shouldClose) notificationManager.closePopup()
+    // Code commented out until we begin auto adding users to NewUI
+    // const useBetaCss = isMascara || firstTime || betaUIState
+    const useBetaCss = isMascara || betaUIState
+
+    let css = useBetaCss ? NewMetaMaskUiCss() : OldMetaMaskUiCss()
+    let deleteInjectedCss = injectCss(css)
+    let newBetaUIState
+
+    store.subscribe(() => {
+      const state = store.getState()
+      newBetaUIState = state.metamask.featureFlags.betaUI
+      if (newBetaUIState !== betaUIState) {
+        deleteInjectedCss()
+        betaUIState = newBetaUIState
+        css = betaUIState ? NewMetaMaskUiCss() : OldMetaMaskUiCss()
+        deleteInjectedCss = injectCss(css)
+      }
+      if (state.appState.shouldClose) notificationManager.closePopup()
+    })
   })
-})
 
 
-function closePopupIfOpen (windowType) {
-  if (windowType !== 'notification') {
-    // should close only chrome popup
-    notificationManager.closePopup()
+  function closePopupIfOpen (windowType) {
+    if (windowType !== 'notification') {
+      // should close only chrome popup
+      notificationManager.closePopup()
+    }
   }
-}
 
-function displayCriticalError (err) {
-  container.innerHTML = '<div class="critical-error">The MetaMask app failed to load: please open and close MetaMask again to restart.</div>'
-  container.style.height = '80px'
-  log.error(err.stack)
-  throw err
+  function displayCriticalError (err) {
+    container.innerHTML = '<div class="critical-error">The MetaMask app failed to load: please open and close MetaMask again to restart.</div>'
+    container.style.height = '80px'
+    log.error(err.stack)
+    throw err
+  }
+
 }