+++ ui/app/components/pending-tx/confirm-send-ether.js
@@ -1,5 +1,5 @@
 const Component = require('react').Component
-const { connect } = require('react-redux')
+const connect = require('../../metamask-connect')
 const h = require('react-hyperscript')
 const inherits = require('util').inherits
 const actions = require('../../actions')
@@ -18,7 +18,6 @@ const {
   isBalanceSufficient,
 } = require('../send/send-utils')
 const GasFeeDisplay = require('../send/gas-fee-display-v2')
-const t = require('../../../i18n')
 const SenderToRecipient = require('../sender-to-recipient')
 const NetworkDisplay = require('../network-display')
 
@@ -205,7 +204,7 @@ ConfirmSendEther.prototype.getData = function () {
     },
     to: {
       address: txParams.to,
-      name: identities[txParams.to] ? identities[txParams.to].name : t('newRecipient'),
+      name: identities[txParams.to] ? identities[txParams.to].name : this.props.t('newRecipient'),
     },
     memo: txParams.memo || '',
     gasFeeInFIAT,
@@ -311,7 +310,7 @@ ConfirmSendEther.prototype.render = function () {
 
         h('div.confirm-screen-rows', [
           h('section.flex-row.flex-center.confirm-screen-row', [
-            h('span.confirm-screen-label.confirm-screen-section-column', [ t('from') ]),
+            h('span.confirm-screen-label.confirm-screen-section-column', [ this.props.t('from') ]),
             h('div.confirm-screen-section-column', [
               h('div.confirm-screen-row-info', fromName),
               h('div.confirm-screen-row-detail', `...${fromAddress.slice(fromAddress.length - 4)}`),
@@ -319,7 +318,7 @@ ConfirmSendEther.prototype.render = function () {
           ]),
 
           h('section.flex-row.flex-center.confirm-screen-row', [
-            h('span.confirm-screen-label.confirm-screen-section-column', [ t('to') ]),
+            h('span.confirm-screen-label.confirm-screen-section-column', [ this.props.t('to') ]),
             h('div.confirm-screen-section-column', [
               h('div.confirm-screen-row-info', toName),
               h('div.confirm-screen-row-detail', `...${toAddress.slice(toAddress.length - 4)}`),
@@ -327,7 +326,7 @@ ConfirmSendEther.prototype.render = function () {
           ]),
 
           h('section.flex-row.flex-center.confirm-screen-row', [
-            h('span.confirm-screen-label.confirm-screen-section-column', [ t('gasFee') ]),
+            h('span.confirm-screen-label.confirm-screen-section-column', [ this.props.t('gasFee') ]),
             h('div.confirm-screen-section-column', [
               h(GasFeeDisplay, {
                 gasTotal: gasTotal || gasFeeInHex,
@@ -345,8 +344,8 @@ ConfirmSendEther.prototype.render = function () {
                 'confirm-screen-section-column': !errors['insufficientFunds'],
               }),
             }, [
-              h('span.confirm-screen-label', [ t('total') + ' ' ]),
-              h('div.confirm-screen-total-box__subtitle', [ t('amountPlusGas') ]),
+              h('span.confirm-screen-label', [ this.props.t('total') + ' ' ]),
+              h('div.confirm-screen-total-box__subtitle', [ this.props.t('amountPlusGas') ]),
             ]),
 
             h('div.confirm-screen-section-column', [
@@ -449,10 +448,10 @@ ConfirmSendEther.prototype.render = function () {
               clearSend()
               this.cancel(event, txMeta)
             },
-          }, t('cancel')),
+          }, this.props.t('cancel')),
 
           // Accept Button
-          h('button.btn-confirm.page-container__footer-button.allcaps', [t('confirm')]),
+          h('button.btn-confirm.page-container__footer-button.allcaps', [this.props.t('confirm')]),
         ]),
       ]),
     ])
@@ -478,9 +477,9 @@ ConfirmSendEther.prototype.onSubmit = function (event) {
   if (valid && this.verifyGasParams() && balanceIsSufficient) {
     this.props.sendTransaction(txMeta, event)
   } else if (!balanceIsSufficient) {
-    updateSendErrors({ insufficientFunds: t('insufficientFunds') })
+    updateSendErrors({ insufficientFunds: this.props.t('insufficientFunds') })
   } else {
-    updateSendErrors({ invalidGasParams: t('invalidGasParams') })
+    updateSendErrors({ invalidGasParams: this.props.t('invalidGasParams') })
     this.setState({ submitting: false })
   }
 }