+++ ui/app/components/sender-to-recipient.js
@@ -1,7 +1,7 @@
 const { Component } = require('react')
 const h = require('react-hyperscript')
+const connect = require('../metamask-connect')
 const PropTypes = require('prop-types')
-const t = require('../../i18n')
 const Identicon = require('./identicon')
 
 class SenderToRecipient extends Component {
@@ -21,7 +21,7 @@ class SenderToRecipient extends Component {
         this.renderRecipientIcon(),
         h(
           '.sender-to-recipient__name.sender-to-recipient__recipient-name',
-          recipientName || t('newContract')
+          recipientName || this.props.t('newContract')
         ),
       ])
     )
@@ -61,6 +61,7 @@ SenderToRecipient.propTypes = {
   senderAddress: PropTypes.string,
   recipientName: PropTypes.string,
   recipientAddress: PropTypes.string,
+  t: PropTypes.func,
 }
 
-module.exports = SenderToRecipient
+module.exports = connect()(SenderToRecipient)