+++ ui/app/components/token-list.js
@@ -3,9 +3,8 @@ const h = require('react-hyperscript')
 const inherits = require('util').inherits
 const TokenTracker = require('eth-token-tracker')
 const TokenCell = require('./token-cell.js')
-const connect = require('react-redux').connect
+const connect = require('../metamask-connect')
 const selectors = require('../selectors')
-const t = require('../../i18n')
 
 function mapStateToProps (state) {
   return {
@@ -43,7 +42,7 @@ TokenList.prototype.render = function () {
   const { tokens, isLoading, error } = state
 
   if (isLoading) {
-    return this.message(t('loadingTokens'))
+    return this.message(this.props.t('loadingTokens'))
   }
 
   if (error) {
@@ -53,7 +52,7 @@ TokenList.prototype.render = function () {
         padding: '80px',
       },
     }, [
-      t('troubleTokenBalances'),
+      this.props.t('troubleTokenBalances'),
       h('span.hotFix', {
         style: {
           color: 'rgba(247, 134, 28, 1)',
@@ -64,7 +63,7 @@ TokenList.prototype.render = function () {
           url: `https://ethplorer.io/address/${userAddress}`,
         })
         },
-      }, t('here')),
+      }, this.props.t('here')),
     ])
   }