+++ ui/app/first-time/init-menu.js
@@ -1,12 +1,11 @@
 const inherits = require('util').inherits
 const EventEmitter = require('events').EventEmitter
 const Component = require('react').Component
-const connect = require('react-redux').connect
+const connect = require('../metamask-connect')
 const h = require('react-hyperscript')
 const Mascot = require('../components/mascot')
 const actions = require('../actions')
 const Tooltip = require('../components/tooltip')
-const t = require('../../i18n')
 const getCaretCoordinates = require('textarea-caret')
 const environmentType = require('../../../app/scripts/lib/environment-type')
 const { OLD_UI_NETWORK_TYPE } = require('../../../app/scripts/config').enums
@@ -60,7 +59,7 @@ InitializeMenuScreen.prototype.renderMenu = function (state) {
           color: '#7F8082',
           marginBottom: 10,
         },
-      }, t('appName')),
+      }, this.props.t('appName')),
 
 
       h('div', [
@@ -70,10 +69,10 @@ InitializeMenuScreen.prototype.renderMenu = function (state) {
             color: '#7F8082',
             display: 'inline',
           },
-        }, t('encryptNewDen')),
+        }, this.props.t('encryptNewDen')),
 
         h(Tooltip, {
-          title: t('denExplainer'),
+          title: this.props.t('denExplainer'),
         }, [
           h('i.fa.fa-question-circle.pointer', {
             style: {
@@ -93,7 +92,7 @@ InitializeMenuScreen.prototype.renderMenu = function (state) {
       h('input.large-input.letter-spacey', {
         type: 'password',
         id: 'password-box',
-        placeholder: t('newPassword'),
+        placeholder: this.props.t('newPassword'),
         onInput: this.inputChanged.bind(this),
         style: {
           width: 260,
@@ -105,7 +104,7 @@ InitializeMenuScreen.prototype.renderMenu = function (state) {
       h('input.large-input.letter-spacey', {
         type: 'password',
         id: 'password-box-confirm',
-        placeholder: t('confirmPassword'),
+        placeholder: this.props.t('confirmPassword'),
         onKeyPress: this.createVaultOnEnter.bind(this),
         onInput: this.inputChanged.bind(this),
         style: {
@@ -120,7 +119,7 @@ InitializeMenuScreen.prototype.renderMenu = function (state) {
         style: {
           margin: 12,
         },
-      }, t('createDen')),
+      }, this.props.t('createDen')),
 
       h('.flex-row.flex-center.flex-grow', [
         h('p.pointer', {
@@ -130,7 +129,7 @@ InitializeMenuScreen.prototype.renderMenu = function (state) {
             color: 'rgb(247, 134, 28)',
             textDecoration: 'underline',
           },
-        }, t('importDen')),
+        }, this.props.t('importDen')),
       ]),
 
       h('.flex-row.flex-center.flex-grow', [
@@ -179,12 +178,12 @@ InitializeMenuScreen.prototype.createNewVaultAndKeychain = function () {
   var passwordConfirm = passwordConfirmBox.value
 
   if (password.length < 8) {
-    this.warning = t('passwordShort')
+    this.warning = this.props.t('passwordShort')
     this.props.dispatch(actions.displayWarning(this.warning))
     return
   }
   if (password !== passwordConfirm) {
-    this.warning = t('passwordMismatch')
+    this.warning = this.props.t('passwordMismatch')
     this.props.dispatch(actions.displayWarning(this.warning))
     return
   }